+++ NEWS.md
@@ -1,3 +1,13 @@
+# sensobol 1.1.3
+
+* Added the function "load_packages" to load (and install in case they are not already in the local R setup) all required libraries in just one call.
+
+# sensobol 1.1.2
+
+* Fixed a bug on sobol_matrices that caused the function to fail when a single parameter was passed into the "params" argument.
+
+* Added a line in the help page of the sobol_indices and sobol_dummy functions informing that the Y argument does not accept NA or NaN values in the model output.
+
 # sensobol 1.1.1
 
 * The CITATION file has changed. Citations to sensobol should now refer to the paper published in the Journal of Statistical Software: